testprocessexecutor.cpp: fixed `bugprone-implicit-widening-of-multiplication-result` clang-tidy warning

+++ test/testprocessexecutor.cpp
@@ -252,7 +252,7 @@ private:
         // for each file: top5 results + overall + empty line
         const std::string output_s = GET_REDIRECT_OUTPUT;
         // for each file: top5 results + overall + empty line
-        TODO_ASSERT_EQUALS((5 + 1 + 1) * 2, 0, cppcheck::count_all_of(output_s, '\n'));
+        TODO_ASSERT_EQUALS(static_cast<long long>(5 + 1 + 1) * 2, 0, cppcheck::count_all_of(output_s, '\n'));
     }
 
     void showtime_top5_summary() {
